[Refactor/Test] MatchResult, 불필요한 메서드 제거 및 테스트 코드 추가
